Adding a debug variable to deleting user-sent commands so that we can toggle this ability on and off. Currently it does not work

+++ DiscordBot/commands/mud/stats.js
@@ -14,8 +14,9 @@ class StatsCommand extends commando.Command {
     async run(message, args) {
         db.getItem(message.member.id, 'players', (data) => this.getPlayer(data, message));
 
-        // delete stat command after saying it
-        message.delete();
+        // delete the user's command if not debugging
+        if (!DEBUG)
+            message.delete();
     }
 
     getPlayer(data, message) {
